Firehose: proof of concept
I'm experimenting with auto-generating nodes for
https://rileyjshaw.com/lab from a variety of data sources. This
project may eventually replace https://rileyjshaw.com.

This is the initial commit, completed quickly as a proof of concept.
There's nothing much to show, but I want to deploy ASAP so I can test
the full pipeline.

So far, everything has worked! Data from a variety of sources is
already appearing on my local server. To reproduce:

- Run `gatsby develop`.
- Open http://localhost:8000/page-2 in a browser.

So far, I'm surfacing data from:

- A hand-curated JSON list of projects I use for
  https://rileyjshaw.com/lab.
- [Dwitter](https://dwitter.net)'s API.
- [My commit--blog](https://rileyjshaw.commit--blog.com)'s Atom feed.

Setting this up was *EASY*, which makes me excited for the future of
this experiment :)