Accessibility: remove "nested link" hack
On my old /lab page, each node was a giant block link. This worked
well for the page, but it's a bit of an accessibility nightmare. Text
selection, scrolling, etc. are obscured behind a floating pseudo
element. The markup worked fine for screen readers, but for sighted
users it's confusing to have a text link within a larger block link.

I'm redesigning the site from scratch, so this is a good opportunity
to come up with a more accessible design. I've watched a few friends
interact with the explorer list, and folks know to click on the
(link styled) headers anyway. I don't think it adds anything to make
the whole box a href target.

Good riddance!