Hotmess: add rjset.css
I wrote this in January before the world exploded. There are so many
lingering changes to my site from January that I cannot place. I’m not
sure what they do, but it looks like a lot of work, so I don’t want to
discard them… but it means I have hardly worked on my site this year.
It’s getting in the way. I like this rjset.css idea, so I’m keeping it.
Layouts etc will be worse for wear after this, but it’s the beginning of
vacumming out the basement.