Release: remove the “beta” banner!!! 🎊
Almost one year ago, I [quietly launched the new version of my site to
rileyjshaw.com](http://rileyjshaw.commit--blog.com/rileyjshaw/rileyjshaw.github.io/acfbf166206e1fcea9f03b6919715eba01e4b2f9).
It was the last day of the decade, and it felt momentous. But since
there were still some glaring bugs and omissions, I decided to hedge the
release behind a “THIS SITE IS IN BETA” banner.

It’s the end of the year again. Though there are _still_ design
problems and omissions, the site is much further along than it was on
initial release. Some highlights:

- [/lab](https://rileyjshaw.com/lab) no-longer crashes Webkit based browsers (lol)
- I added _bunch_ of new scrapers. Most recently: SoundCloud, Vimeo, and
  YouTube.
- The design is still uh… getting there… but it’s at least simpler and
  more consistent now.
- I test-drove the blog and added some new features (like MDX!)
- I put a ton of meat on last year’s skelly bones. From
  [/about](https://rileyjshaw.com/about) to [/lab](https://rileyjshaw.com/lab)
  to adding about 3x more quotes, this site is more substantial and
  chaotic than ever before.

There have been about 130 commits between the beta release and this one.
It’s not much for a whole year, but fixing up my personal website felt
pretty small and self-indulgent vs. everything else I was focusing on.
And 130 is still plenty of work to call this beta version un-betafied!

I was initially going to remove the banner on New Year’s Eve, but I’d
rather do it early and spend the rest of the year offline. That feels
like the biggest improvement in this whole feature list.