Hotmess: add a tags curation tool
Tags have been a part of this site – somewhat – since the beginning. But
most content was untagged, and the ratio has gotten worse over time.
This tool provides a quick way for me to apply the standard list of list
of tags to new content, and add tags. It’s the first of a few similar
curation tools I have planned, but hopefully I resist the urge to build
them all at once, and my next few curation commits are just from using
this tool :)