Tooling: format generated JSON files on save & stop prettier
Prettier takes a while to chug through the `project-scraper/_generated`
JSONs, and basically does the same thing as calling
JSON.stringify(content, null, '\t') would have anyway! This saves a
bunch of time on lint, and also… hopefully… means that the diffs for
`scraped-projects-raw.json` are WAAAAY smaller.

And speaking of diffs, I replaced JSON.stringify with
json-stable-stringify in a few places. That should help a _lot_ with
scraped content.